src/third_party/blink/web_tests/dom/domparsing/xmlserializer-attribute-special-namespaces-expected.txt

Tests the serialization of special XML namespaces on attributes, as reported in bug 395950.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xmlDocument.querySelector("inner1").getAttributeNS("http://www.w3.org/XML/1998/namespace", "attr") is "value"
PASS xmlDocument.querySelector("inner2").getAttributeNS("http://www.w3.org/XML/1998/namespace", "attr") is "value"
PASS xmlDocument.querySelector("inner2").getAttribute("xml:attr") is "value"
PASS parsedDoc.querySelector("inner1").getAttributeNS("http://www.w3.org/XML/1998/namespace", "attr") is "value"
PASS parsedDoc.querySelector("inner2").getAttributeNS("http://www.w3.org/XML/1998/namespace", "attr") is "value"
PASS parsedDoc.querySelector("inner2").getAttribute("xml:attr") is "value"
PASS xmlString.indexOf("xmlns:xml") is -1
PASS successfullyParsed is true

TEST COMPLETE
